Structure complete, ready for review.
All the files are included here; every section we each worked on has a folder and inside there is a child folder "HeaderFiles" which contains the Header Files for the parent folder. I did not include any of the diagrams inside the UML diagram, please do so. Don't forget to create pdf ;)